Kernel-repository parameter added, version-tag parameter changed to head_sha parameter

+++ scripts/analyse-kernel.sh
@@ -1,21 +1,49 @@
 #!/bin/bash
 
 help() {
-	echo "Usage: $0 [-v] [-c] [-i]"
+	echo "Usage: $0 [-r] [-s] [-c] [-i]"
 	echo "Or you can directly provide a analysisconfig file with [--configfile]"
-	echo "  -v,  kernel-version"
+	echo "  -r,  kernel-repository"
+	echo "  -s,  kernel-sha"
 	echo "  -c,  kernel-configuration"
 	echo "  -i,  inferconfig-file location"
 	echo "  --configfile, analysisconfig file for build"
-	echo "Example: $0 -v 4.16 -c defconfig -i [absolute-path-to-file]"
+	echo "Example: $0 -r linux-stable -s d8a5b80568a9 -c defconfig -i [absolute-path-to-file]"
 	echo "Example with analysisconfig file: $0 --configfile [absolute-path-to-file]"
 	echo "Before use it, please be sure you set KERNEL_SRC_BASE variable correctly"
 	exit 1
 }
 assign_default_values() {
-	KERNEL_VERSION="v4.15"
+	KERNEL_REPOSITORY="$KERNEL_SRC_BASE/stable/linux-stable"
+	KERNEL_VERSION="d8a5b80568a9"
 	KERNEL_CONFIG="defconfig"
 	INFERCONFIG_LOCATION="$SCRIPTS_DIRECTORY/files/inferconfig"
+}
+set_kernel_repository() {
+	if [ ! -z "$1" ]; then
+		case $1 in
+			torvalds | stable | next )
+			if [ "$1" == "torvalds" ]; then
+				echo "Torvalds!"
+				KERNEL_REPOSITORY="$KERNEL_SRC_BASE/torvalds/linux"
+			elif [ "$1" == "stable" ]; then
+				echo "Stable!"
+				KERNEL_REPOSITORY="$KERNEL_SRC_BASE/stable/linux-stable"
+			elif [ "$1" == "next" ]; then
+				echo "Next!"
+				KERNEL_REPOSITORY="$KERNEL_SRC_BASE/next/linux-next"
+			else
+				echo "Invalid Kernel Repository Parameter!"
+				echo "You can pass -> torvalds | stable | next"
+				exit 1;
+			fi
+		esac
+	else
+		echo "You provided an empty KERNEL_REPOSITORY parameter!"
+		echo "Script will use linux-stable repository for this run!"
+		KERNEL_REPOSITORY="$KERNEL_SRC_BASE/stable/linux-stable"
+		exit 1;
+	fi
 }
 set_kernel_config() {
 	if [ ! -z "$1" ]; then
@@ -59,7 +87,7 @@ elif [ ! -d "$KERNEL_SRC_BASE" ]; then
 	exit 1
 fi
 SCRIPTS_DIRECTORY="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-KERNEL_STABLE_DIRECTORY="$KERNEL_SRC_BASE/stable/linux-stable"
+#KERNEL_STABLE_DIRECTORY="$KERNEL_SRC_BASE/stable/linux-stable"
 # User need help or called without params
 if [[ "$1" == "-h" || "$1" == "--help" ]]; then
 	help
@@ -67,19 +95,26 @@ if [[ "$1" == "-h" || "$1" == "--help" ]]; then
 fi
 assign_default_values
 while [[ "$#" > 0 ]]; do case $1 in
-  -v) KERNEL_VERSION="$2"; shift; shift;;
+  -r) set_kernel_repository "$2"; shift; shift;;
+  -s) KERNEL_HEAD_SHA="$2"; shift; shift;;
   -c) set_kernel_config "$2"; shift; shift;;
   -i) set_inferconfig_file_location "$2"; shift; shift;;
   --configfile) read_and_set_variables_from_analysisconfig "$2"; shift; shift;;
   *) help; shift; shift; exit 1;;
 esac; done
-cd $KERNEL_STABLE_DIRECTORY
-git clean -f -d
-git checkout .
-git checkout $KERNEL_VERSION
-cp "$INFERCONFIG_LOCATION" "$KERNEL_STABLE_DIRECTORY/.inferconfig"
+if [ ! -d "$KERNEL_REPOSITORY" ]; then
+	echo "Path $KERNEL_REPOSITORY doesn't point to a directory"
+	echo "Please fix it and then run script"
+	exit 1
+fi
+echo "KERNEL_REPOSITORY is = $KERNEL_REPOSITORY"
+cd $KERNEL_REPOSITORY
+#git clean -f -d
+#git checkout .
+git checkout $KERNEL_HEAD_SHA
+cp "$INFERCONFIG_LOCATION" "$KERNEL_REPOSITORY/.inferconfig"
 git apply "$SCRIPTS_DIRECTORY/files/0001-Set-default-CC-to-Clang-from-Makefile.patch"
 DOCKER_NAME="kernel-analysis"
-docker run -v "$KERNEL_STABLE_DIRECTORY:/linux/" --interactive --tty $DOCKER_NAME \
+docker run -v "$KERNEL_REPOSITORY:/linux/" --interactive --tty $DOCKER_NAME \
 /bin/sh -c "cd linux && make clean && make $KERNEL_CONFIG && infer capture -- make -j40 && git checkout . && git checkout master && rm .inferconfig"
